+++ src/problem_1.rs
@@ -1,5 +1,5 @@
 //! Multiples of 3 and 5
 
-pub fn solve() -> i32 {
-    (3..1000).filter(|n| n % 3 == 0 || n % 5 == 0).sum::<i32>()
+pub fn solve() -> u64 {
+    (3..1000).filter(|n| n % 3 == 0 || n % 5 == 0).sum::<u64>()
 }
